SPRINT 22 часть 13 Исправил ещё одно некритичное замечание про дублирование кода - пошёл по пути Extensions. Поправил разметку BottomSheetLayout Добавил в тост название плейлиста, в который добавляем TRACK
