+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2025-08-01T14:48:47.251135200Z">
+        <DropdownSelection timestamp="2025-08-01T15:21:56.793294200Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\Fantasma\.android\avd\Pixel_3_XL_API_34.avd" />
+              <DeviceId pluginId="PhysicalDevice" identifier="serial=R5CRA02RAXE" />
             </handle>
           </Target>
         </DropdownSelection>
