+++ app/src/main/java/com/practicum/playlistmaker/media/fragments/playlists/ui/view/PlaylistAdapter.kt
@@ -7,8 +7,10 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
 import com.practicum.playlistmaker.R
 import com.practicum.playlistmaker.media.fragments.playlists.ui.PlaylistUi
+import com.practicum.playlistmaker.util.getTrackCountText
 
 class PlaylistAdapter : RecyclerView.Adapter<PlaylistAdapter.PlaylistViewHolder>() {
 
@@ -39,28 +41,21 @@ class PlaylistAdapter : RecyclerView.Adapter<PlaylistAdapter.PlaylistViewHolder>
 
         fun bind(playlist: PlaylistUi) {
             titleText.text = playlist.name
-            descriptionText.text = "${getTrackCountText(playlist.trackCount)}"
+            descriptionText.text = getTrackCountText(playlist.trackCount)
 
             if (!playlist.coverPath.isNullOrEmpty()) {
-                coverImage.setImageURI(Uri.parse(playlist.coverPath))
+                Glide.with(itemView)
+                    .load(Uri.parse(playlist.coverPath))
+                    .placeholder(R.drawable.playlist_cover_placeholder)
+                    .centerCrop()
+                    .into(coverImage)
             } else {
                 coverImage.setImageResource(R.drawable.playlist_cover_placeholder)
             }
+
         }
 
-        fun getTrackCountText(count: Int): String {
-            val remainder10 = count % 10
-            val remainder100 = count % 100
 
-            val word = when {
-                remainder100 in 11..14 -> "треков"
-                remainder10 == 1 -> "трек"
-                remainder10 in 2..4 -> "трека"
-                else -> "треков"
-            }
-
-            return "$count $word"
-        }
     }
 }
