+++ app/src/main/java/com/practicum/playlistmaker/media/fragments/playlists/ui/viewmodel/PlaylistsViewModel.kt
@@ -47,5 +47,4 @@ class PlaylistsViewModel(
                 }
         }
     }
-
 }
