+++ app/src/main/java/com/practicum/playlistmaker/util/Extensions.kt
@@ -1,0 +1,15 @@
+package com.practicum.playlistmaker.util
+
+    fun getTrackCountText(count: Int): String {
+        val remainder10 = count % 10
+        val remainder100 = count % 100
+
+        val word = when {
+            remainder100 in 11..14 -> "треков"
+            remainder10 == 1 -> "трек"
+            remainder10 in 2..4 -> "трека"
+            else -> "треков"
+        }
+
+        return "$count $word"
+    }
