+++ app/src/main/res/layout/bottom_sheet_add_to_playlist.xml
@@ -2,22 +2,22 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/bottom_sheet_content"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
+    android:layout_width="match_parent"
+    android:layout_height="505dp"
     android:orientation="vertical"
     android:padding="16dp"
-    android:background="?attr/colorPrimary">
+    android:background="@drawable/bottom_sheet_handle"
+    android:gravity="center_horizontal">
 
     <View
         android:layout_width="40dp"
         android:layout_height="4dp"
-        android:layout_gravity="center"
-        android:background="@drawable/bottom_sheet_handle"
+        android:layout_gravity="center_horizontal"
         android:layout_marginBottom="16dp"/>
 
     <TextView
         android:id="@+id/title_add_to_playlist"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="Добавить в плейлист"
         android:gravity="center_horizontal"
@@ -39,7 +39,7 @@
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_add_to_playlist"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
+        android:layout_height="wrap_content"
         android:layout_weight="1"
         android:clipToPadding="false"
         android:paddingBottom="8dp"/>
